requirements.txt
mysql-connector-python
tkcalendar
customtkinter